+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 - `CodeGenerator`, `Validation`, templates: optimised decimal comparisons
 - `CodeGenerator`, `Validation`, templates: optimised integer and long range checks
+- `CodeGenerator`: bug fix - no longer outputs checks for strings with format causing them to be other class
 
 ## [0.28] - 2021-03-23
 ### Changed